api/pyproject.toml
[project]
name = "coherence-api"
version = "0.1.0"
description = "Coherence Network API"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]
"app/main.py" = ["E402"]
"app/routers/agent.py" = ["E402"]
"tests/test_project_manager*.py" = ["E402", "F401"]
"tests/conftest.py" = ["F401"]
